Merge branch 'sctp_command_queue'
David Laight says:

====================
net: sctp: Optimisations to sctp command queue code

These 3 patches optimise the code that processes sctp's command queue.
(A list of 'tasks' to be performed after the rest of the chunk processing.)

1) Inline all the functions from command.c
2) Remove the memset() calls used to zero a word-sized union.
3) Use pointers instead of array indexes.

The combined changes reduce the code size (amd64) by a few kb.

I'm not 100% convinced that the zeroing done in patch 2 is needed at all.
On BE systems it is likely to generate more code than on LE ones.
In fact it might be best to change the union to only contain 'long' sized
items.

Changes for v2:
	- Add some missing initialisers in patch 2/3 and delete them in 3/3.
	- Modify the commit message for 2/3 to point out that the union
	  shouldn't need to be zeroed, but the patches aren't intended to
	  change the behaviour even if the code is buggy.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 90fb5679e568b11b89e02b87f7f4fe00c7589ce0 ]